refactor(Machine): Separate the class into Machine and Utils classes
Machine 클래스 내 getString과 printResult 함수를 Utils 클래스로 분리

Application.java 내 정의된 Machine 클래스를 분리하여 Machine.java와
Utils.java에 정의.
